requirements.txt
ago==0.0.93
async-generator==1.10
attrs==21.4.0
Automat==20.2.0
awscli==1.22.49
beautifulsoup4==4.10.0
botocore==1.23.49
bs4==0.0.1
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
chromedriver-binary==99.0.4844.17.0
click==8.0.3
click-config-file==0.6.0
click-plugins==1.1.1
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
docutils==0.15.2
dotmap==1.3.26
elasticsearch==7.17.0
fake-useragent==0.1.11
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.4.2
fix-yahoo-finance==0.1.30
Flask==2.0.2
Flask-Cors==3.0.10
h11==0.13.0
h2==3.2.0
hjson==3.0.2
hpack==3.0.0
humanize==3.14.0
hurry.filesize==0.9
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
jieba3k==0.35.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
langdetect==1.0.9
lxml==4.7.1
MarkupSafe==2.0.1
multitasking==0.0.10
news-fetch==0.2.8
news-please==1.5.21
newsapi-python==0.2.6
newspaper3k==0.2.8
nltk==3.6.7
numpy==1.22.2
oauthlib==3.2.0
outcome==1.1.0
pandas==1.4.0
parsel==1.6.0
Pillow==9.0.1
plac==1.3.4
priority==1.3.0
Protego==0.1.16
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
PyMySQL==1.0.2
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
PyYAML==5.4.1
queuelib==1.6.2
readability-lxml==0.8.1
regex==2022.1.18
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.5.1
Scrapy==2.5.1
selenium==4.1.0
service-identity==21.1.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
tinysegmenter==0.3
tldextract==3.1.2
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
twarc==2.9.2
tweepy==4.5.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typing_extensions==4.0.1
Unidecode==1.3.2
urllib3==1.26.8
w3lib==1.22.0
warcio==1.7.4
Werkzeug==2.0.2
wsproto==1.0.0
yfinance==0.1.70
zope.interface==5.4.0